requirements.txt
blinker==1.4
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
d20
dnspython==1.16.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
newrelic==5.2.0.127
numexpr==2.7.0
numpy==1.18.1
PyJWT==1.7.1
pymongo==3.9.0
redis==3.3.8
requests==2.22.0
sentry-sdk[flask]==0.12.3
six==1.12.0
urllib3==1.25.6
Werkzeug==0.16.0
